# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Talent Verification & Reward System API
  description: Dokumentasi API untuk MVP "Talent Verification & Reward System"
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /submissions:
    post:
      summary: Membuat pengajuan sertifikat atau portofolio baru
      tags:
        - Submissions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - category
                - subType
                - title
                - description
                - evidence
              properties:
                category:
                  type: string
                  enum: [sertifikat, portofolio]
                subType:
                  type: string
                  enum: [lokal, regional, nasional, internasional, personal, freelance, industri, juara_kompetisi]
                title:
                  type: string
                description:
                  type: string
                evidence:
                  type: string
                  format: binary
      responses:
        '201':
          description: Pengajuan berhasil dibuat

    get:
      summary: Menampilkan semua pengajuan milik mahasiswa yang sedang login
      tags:
        - Submissions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daftar pengajuan berhasil diambil
  /submissions/{id}:
    get:
      summary: Menampilkan detail pengajuan berdasarkan ID
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail pengajuan berhasil diambil

    put:
      summary: Mengajukan ulang pengajuan yang ditolak (status rejected)
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                subType:
                  type: string
                evidence:
                  type: string
                  format: binary
      responses:
        '200':
          description: Pengajuan berhasil dikirim ulang

    delete:
      summary: Menghapus pengajuan (hanya jika berstatus pending/rejected)
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pengajuan berhasil dihapus
...
  /rewards:
    get:
      summary: Menampilkan daftar reward yang aktif
      tags:
        - Rewards
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daftar reward berhasil diambil

    post:
      summary: Membuat reward baru (Khusus Admin)
      tags:
        - Rewards
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - pointRequired
                - stock
              properties:
                title:
                  type: string
                description:
                  type: string
                pointRequired:
                  type: integer
                stock:
                  type: integer
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Reward berhasil dibuat
  /rewards/{id}:
    put:
      summary: Mengupdate data reward (Khusus Admin)
      tags:
        - Rewards
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                pointRequired:
                  type: integer
                stock:
                  type: integer
                  isActive:
                    type: string
                  image:
                    type: string
                    format: binary
      responses:
        '200':
          description: Reward berhasil diupdate
  /rewards/{id}/claim:
    post:
      summary: Mengklaim reward (Khusus Mahasiswa)
      tags:
        - Rewards
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Klaim reward sukses dan poin didecrement
  /rewards/history:
    get:
      summary: Menampilkan riwayat klaim reward milik mahasiswa
      tags:
        - Rewards
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Riwayat klaim berhasil diambil
  /rewards/claims:
    get:
      summary: Menampilkan daftar semua klaim mahasiswa (Khusus Admin)
      tags:
        - Rewards
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daftar klaim berhasil diambil
  /admin/submissions:
    get:
      summary: Menampilkan antrian verifikasi pengajuan (Khusus Admin)
      tags:
        - Admin Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        '200':
          description: Daftar antrian berhasil diambil
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT